feat: Update CategoryHeaderPopupConfig and CategoryHeader components, add some css to improve ux
This commit updates the `CategoryHeaderPopupConfig` file to add a new type `CategoryItems` for future use. It also updates the `CategoryHeader` component by adding a new class to the `DialogTrigger` element and modifying the `DialogContent` element to include absolute positioning and a larger maximum width. These changes improve the functionality and styling of the category header popup.